# docs/openapi/gurgel-core.yaml
openapi: 3.0.3
info:
  title: Gurgel Team API
  version: 1.0.0
  description: |
    Contrato HTTP núcleo — Fase 4 (v1 completo — Fase 4).
    Schemas Zod canônicos em `lib/contracts/api/v1/`.
servers:
  - url: /api/v1
    description: Next.js API routes (alvo Fase 5)
  - url: https://api.gurgelteam.com.br/v1
    description: Produção (futuro)
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    sessionCookie:
      type: apiKey
      in: cookie
      name: session
  schemas:
    ApiError:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          enum:
            - VALIDATION_ERROR
            - UNAUTHORIZED
            - FORBIDDEN
            - NOT_FOUND
            - CONFLICT
            - BUSINESS_RULE
            - RATE_LIMITED
            - SERVICE_UNAVAILABLE
            - INTERNAL_ERROR
        message:
          type: string
        httpStatus:
          type: integer
        details:
          type: object
    ErrorResponse:
      type: object
      required: [success, error]
      properties:
        success:
          type: boolean
          enum: [false]
        error:
          $ref: "#/components/schemas/ApiError"
        message:
          type: string
...
paths:
...
  /inventory/parts/{partId}:
    patch:
      tags: [Inventory]
      summary: Atualizar peça
      security:
        - bearerAuth: []
      parameters:
        - name: partId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: OK
        "401":
          description: Não autenticado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Não encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...